requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.2.0
contourpy==1.1.1
cssselect2==0.7.0
cycler==0.12.1
Django==4.2.5
django-bootstrap-datepicker-plus==5.0.4
django-debug-toolbar==4.2.0
django-excel==0.0.10
django-jazzmin @ git+https://github.com/farridav/django-jazzmin.git@586e81f204033add57b67f657e43681801356a9a
django-mathfilters==1.0.0
django-nepali-datetime-field==0.6
et-xmlfile==1.1.0
fonttools==4.47.2
html5lib==1.1
idna==3.4
importlib-resources==6.1.1
kiwisolver==1.4.5
lml==0.1.0
matplotlib==3.7.4
nepali==1.1.1
nepali-date-converter==0.0.1
nepali-datetime==1.0.8.1
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
# pkg_resources==0.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydyf==0.10.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pyparsing==3.1.1
pyphen==0.15.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==6.6.0
texttable==1.7.0
tinycss2==1.3.0
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.0.5
uuid==1.30
weasyprint==61.2
webencodings==0.5.1
zipp==3.17.0
zopfli==0.2.3
django-crispy-forms==2.3
crispy-bootstrap5==2024.10  # Include crispy-bootstrap5 dependency


